\name{FancyPack.R.LiU.4-package}
\alias{FancyPack.R.LiU.4-package}
\alias{FancyPack.R.LiU.4}
\docType{package}
\title{
\packageTitle{FancyPack.R.LiU.4}
}
\description{
\packageDescription{FancyPack.R.LiU.4}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FancyPack.R.LiU.4}
\packageIndices{FancyPack.R.LiU.4}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{FancyPack.R.LiU.4}

Maintainer: \packageMaintainer{FancyPack.R.LiU.4}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
